move C specific FileUtil functions to CUtil
